#7 Model/SeedData.cs:Create a seed data class to populate the database
This is used to generate random test data when the application is running.  Data is read from the .json file and inserted into the parent database tables.

Then 20 repair records are created with some having both date opened and date completed while others have only date opened entries.

Then 100 items are created with randome serial numbers which are assigned to 100 repair items.

For each repair item, it will get a random number of faults and resolutions assigned to it.