#7 Program.cs:Execute the SeedData class when the application starts
If there are no records in the database when the application starts, it will automatically be populated with seed data.

With .net core 3.0 you have to put your data seeding in the Program class.  Also a try catch is wrapped around it to display any errors during the data seeding.