#36 Restrict access to the CustomerController
With the authorize attribute at the top of the class, only users in the group will have access to the methods.

The AllowAnonymous attribute means anyone can use the specific method.

With HttpContext.User.IsInRole() you can restrict certain output for data depending on what the role is for the user.