#36 Create the IdentityDataContext class
The ASP.NET Core Identity system depends on Entity Framework Core for access to its data, which means that a database context class is required.

The IdentityDbContext base class provides the properties through which the data is accessed, and creating a subclass allows the Identity features to be customized and extended.