#36 Adding Identity to the Application Configuration
This will set up ASP.NET Core Identity so that it is integrated with the rest of the ASP.NET Core application.

The AddDbContext method called in ConfigureServices registers the Identity database context class with Entity Framework Core, while the AddIdentity method tells Identity to use the default classes to represent users and roles and to store the data through the context class.